# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "People statistic",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/people": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Получить список людей с фильтрацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по имени",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по фамилии",
                        "name": "last_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по отчеству",
                        "name": "patronymic",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по возрасту",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по национальности",
                        "name": "nationalize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по полу",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Лимит записей",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PeopleFullDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Создать нового человека",
                "parameters": [
                    {
                        "description": "Данные для создания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePeopleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданного человека",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/people/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Получить человека по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID человека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PeopleFullDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Обновить данные человека",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID человека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновляемые данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PeopleDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "People"
                ],
                "summary": "Удалить человека",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID человека",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreatePeopleDTO": {
            "type": "object",
            "required": [
                "last_name",
                "name"
            ],
            "properties": {
                "last_name": {
                    "type": "string",
                    "minLength": 2
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "patronymic": {
                    "type": "string"
                }
            }
        },
        "dto.PeopleDTO": {
            "type": "object",
            "required": [
                "age",
                "gender",
                "last_name",
                "name",
                "nationalize"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nationalize": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                }
            }
        },
        "dto.PeopleFullDTO": {
            "type": "object",
            "required": [
                "age",
                "gender",
                "last_name",
                "name",
                "nationalize"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nationalize": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                }
            }
        }
    }
}